refactor: move files to components based structure
