+++ src/helpers/__mocks__/localStorage.js
